Moved some code to config_utility
Removed a duped var vs_editions as it is already in visualstudio.py
